+++ app/vs_gst_qmlglsink/gstqmlglsinkstream.cpp
@@ -1,4 +1,4 @@
-#include "gstvideostream.h"
+#include "gstqmlglsinkstream.h"
 
 #include <QtQuick>
 #include <sstream>
